Rich Text: Move RichTextShortcut and RichTextInputEvent into @wordpress/rich-text
These format helper components and their contexts have no block-editor
specific logic, so house them in the rich-text package where any rich
text field owner (not just the block-editor canvas) can provide the
contexts. The block-editor files re-export them from the rich-text
private APIs for back-compat, since format-library and third-party
formats import them from @wordpress/block-editor.

The move also fixes the react-hooks/refs suppression in both
components by keeping the latest callback in a ref updated inside an
effect instead of during render.

Extracted from #78471 per review feedback.
