+++ packages/rich-text/src/private-apis.js
@@ -3,6 +3,9 @@
  */
 import { lock } from './lock-unlock';
 import { useRichText } from './hook';
+import { keyboardShortcutContext, inputEventContext } from './contexts';
+import { RichTextShortcut } from './keyboard-shortcut';
+import { RichTextInputEvent } from './input-event';
 
 /**
  * Private @wordpress/rich-text APIs.
@@ -10,4 +13,8 @@ import { useRichText } from './hook';
 export const privateApis = {};
 lock( privateApis, {
 	useRichText,
+	keyboardShortcutContext,
+	inputEventContext,
+	RichTextShortcut,
+	RichTextInputEvent,
 } );
